+++ modules/core/CHANGELOG.md
@@ -23,6 +23,11 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Introducing the Selection API that will be used by components such as check boxes, radio buttons, switches and such.
   However, the API has been developed to be as general purpose as possible, meaning that it can be used with anything
   implementing the needed interface
+- Added tests for the new Selection API
+
+## Changed
+
+- Moved CSSFragment from components
 
 ## [11.3.1] - 11-04-2023